# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "jobApplier API",
  "version": "0.1.0"
 },
 "paths": {
  "/api/auth/login": {
   "post": {
    "summary": "Login",
    "operationId": "login_api_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/controls/pause": {
   "post": {
    "summary": "Pause",
    "operationId": "pause_api_controls_pause_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PauseIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/controls/resume": {
   "post": {
    "summary": "Resume",
    "operationId": "resume_api_controls_resume_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PauseIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/controls/retention": {
   "put": {
    "summary": "Put Retention",
    "operationId": "put_retention_api_controls_retention_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RetentionIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/audit": {
   "get": {
    "summary": "List Audit",
    "operationId": "list_audit_api_audit_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "before_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Before Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/privacy/delete": {
   "post": {
    "summary": "Delete Everything",
    "operationId": "delete_everything_api_privacy_delete_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeleteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "DeleteIn": {
    "properties": {
     "confirm": {
      "type": "string",
      "title": "Confirm"
     }
    },
    "type": "object",
    "required": [
     "confirm"
    ],
    "title": "DeleteIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginIn": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "title": "Username"
     },
     "password": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Password"
     },
     "totp": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 10
       },
       {
        "type": "null"
       }
      ],
      "title": "Totp"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "LoginIn"
   },
...
   "PauseIn": {
    "properties": {
     "reason": {
      "type": "string",
      "maxLength": 500,
      "title": "Reason",
      "default": ""
     }
    },
    "type": "object",
    "title": "PauseIn"
   },
...
   "RetentionIn": {
    "properties": {
     "expired_job_days": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Expired Job Days"
     },
     "audit_days": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 30.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Audit Days"
     },
     "connector_run_days": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Connector Run Days"
     },
     "rejected_application_days": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Rejected Application Days"
     }
    },
    "type": "object",
    "title": "RetentionIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}